// src/types/row.ts
export interface Row {
   child: Row[];
   id: number;
   rowName: string;
   equipmentCosts: number;
   estimatedProfit: number;
   machineOperatorSalary: number;
   mainCosts: number;
   materials: number;
   mimExploitation: number;
   overheads: number;
   salary: number;
   supportCosts: number;
   total: number;
}